requirements.txt
colorama==0.4.4
requests==2.25.1
speedtest-cli==2.1.3
